refactor: separate complete button
